VouchUser to utilize TicketHandler for ticket creation; remove redundant ticket creation logic from VouchUser. Update VouchHandler to notify relevant roles upon vouch approval.
